backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
opencv-contrib-python-headless==4.10.0.84
rembg==2.0.72
onnxruntime>=1.18.0
Pillow>=10.0.0
numpy>=1.26.0
